Created weathertool library & added arguent and flags to the program
